@@ -19,7 +19,7 @@
   },
   "scripts": {
     "prepare": "node -e \"try { require('husky').install() } catch (e) {if (e.code !== 'MODULE_NOT_FOUND') throw e}\"",
-    "dev": "vite",
+    "start": "vite",
     "build": "rm -rf dist && npm run lint:scripts && npm run lint:styles && vite build",
     "version": "changeset version",
     "release": "changeset release",
@@ -61,7 +61,6 @@
   "author": "Andrew Rubin <andrew@wethecollective.com> (https://wethecollective.com)",
   "license": "MIT",
   "devDependencies": {
-    "@changesets/changelog-github": "~0.4.8",
     "@types/node": "~20.9.4",
     "@types/react": "~18.2.38",
     "@types/react-dom": "~18.2.16",
@@ -80,8 +79,8 @@
     "postcss": "^8.4.23",
     "postcss-scss": "^4.0.6",
     "prettier": "~2.8.8",
-    "react": "17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "sass": "^1.62.1",
     "stylelint": "^15.6.2",
     "stylelint-config-recommended": "^12.0.0",
@@ -94,11 +93,12 @@
     "vite-plugin-dts": "~3.6.3"
   },
   "peerDependencies": {
-    "react": "17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "17 - 18",
+    "react-dom": "17 - 18"
   },
   "dependencies": {
+    "@changesets/changelog-github": "~0.4.8",
     "@changesets/cli": "~2.26.2",
-    "@wethegit/react-hooks": "^1.0.0"
+    "@wethegit/react-hooks": "^2.0.0"
   }
 }
